fix: client fatal
NULL tile entity reference in Collector sync packet! Please report to dev!

+++ src/main/java/moze_intel/projecte/network/packets/CollectorSyncPKT.java
@@ -30,7 +30,7 @@ public class CollectorSyncPKT implements IMessage {
 
     @Override
     public void fromBytes(ByteBuf buf) {
-        displayEmc = buf.readInt();
+        displayEmc = buf.readLong();
         displayKleinCharge = buf.readDouble();
         x = buf.readInt();
         y = buf.readInt();
